requirements.txt
streamlit
google-genai
pydantic
opencv-python
torch
transformers
faiss-cpu
Pillow
openai
ollama
